docs: rename in docs
