+++ packages/angular-test-app/src/preview-examples/content-header-no-back.html
@@ -1,4 +1,4 @@
-<ix-page-header
+<ix-content-header
   variant="Secondary"
   headerTitle="content title"
   headerSubtitle="subtitle content"
@@ -12,4 +12,4 @@
   <ix-icon-button icon="context-menu" ghost="true" variant="Primary">
     Button3
   </ix-icon-button>
-</ix-page-header>
+</ix-content-header>
