+++ packages/angular-test-app/src/preview-examples/content-header.html
@@ -1,4 +1,4 @@
-<ix-page-header
+<ix-content-header
   hasBackButton="true"
   headerTitle="content title"
   headerSubtitle="subtitle content"
@@ -6,4 +6,4 @@
   <ix-button ghost="true">Button1</ix-button>
   <ix-button ghost="true">Button2</ix-button>
   <ix-button ghost="true">Button3</ix-button>
-</ix-page-header>
+</ix-content-header>
