+++ packages/html-test-app/src/preview-examples/content-header-no-back.html
@@ -11,7 +11,7 @@ SPDX-License-Identifier: MIT
   </head>
   <body class="theme-brand-dark">
     <!-- Preview code -->
-    <ix-page-header
+    <ix-content-header
       variant="Secondary"
       header-title="Content title"
       header-subtitle="Subtitle"
@@ -25,7 +25,7 @@ SPDX-License-Identifier: MIT
       <ix-icon-button icon="context-menu" ghost="true" variant="Primary">
         Button3
       </ix-icon-button>
-    </ix-page-header>
+    </ix-content-header>
     <!-- Preview code -->
     <script type="module" src="./init.js"></script>
   </body>
