+++ packages/html-test-app/src/preview-examples/content-header.html
@@ -11,7 +11,7 @@ SPDX-License-Identifier: MIT
   </head>
   <body class="theme-brand-dark">
     <!-- Preview code -->
-    <ix-page-header
+    <ix-content-header
       has-back-button="true"
       header-title="Content title"
       header-subtitle="Subtitle"
@@ -19,7 +19,7 @@ SPDX-License-Identifier: MIT
       <ix-button ghost="true">Button1</ix-button>
       <ix-button ghost="true">Button2</ix-button>
       <ix-button ghost="true">Button3</ix-button>
-    </ix-page-header>
+    </ix-content-header>
     <!-- Preview code -->
     <script type="module" src="./init.js"></script>
   </body>
