+++ packages/react-test-app/src/main.tsx
@@ -34,6 +34,7 @@ import Buttons from './preview-examples/buttons';
 import Checkbox from './preview-examples/checkbox';
 import CheckboxIndeterminate from './preview-examples/checkbox-indeterminate';
 import Chip from './preview-examples/chip';
+import PageHeader from './preview-examples/content-header';
 import Datepicker from './preview-examples/datepicker';
 import DatepickerRange from './preview-examples/datepicker-range';
 import Datetimepicker from './preview-examples/datetimepicker';
@@ -59,7 +60,6 @@ import MapNavigation from './preview-examples/map-navigation';
 import MapNavigationOverlay from './preview-examples/map-navigation-overlay';
 import MessageBar from './preview-examples/message-bar';
 import ModalExample from './preview-examples/modal';
-import PageHeader from './preview-examples/page-header';
 import Pill from './preview-examples/pill';
 import PopoverNews from './preview-examples/popover-news';
 import RadioButton from './preview-examples/radio-button';
